refactor: separate response logic to file
