+++ internal/handler/maincateg/maincateg.go
@@ -81,36 +81,12 @@ func (m *MainCategHandler) GetAllMainCateg(w http.ResponseWriter, r *http.Reques
 		return
 	}
 
-	// TODO: refactor the following logic
-	type icon struct {
-		ID  int64  `json:"id"`
-		URL string `json:"url"`
-	}
-	type resp struct {
-		ID   int64  `json:"id"`
-		Name string `json:"name"`
-		Type string `json:"type"`
-		Icon icon   `json:"icon"`
-	}
-
-	var respCategs []*resp
-	for _, categ := range categs {
-		respCategs = append(respCategs, &resp{
-			ID:   categ.ID,
-			Name: categ.Name,
-			Type: categ.Type.String(),
-			Icon: icon{
-				ID:  categ.Icon.ID,
-				URL: categ.Icon.URL,
-			},
-		})
-	}
-
+	resp := cvtToGetAllMainCategResp(categs)
 	respData := map[string]interface{}{
-		"categories": respCategs,
+		"categories": resp.Categories,
 	}
+
 	if err := jsonutil.WriteJSON(w, http.StatusOK, respData, nil); err != nil {
-		logger.Error("jsonutil.WriteJSON failed", "package", "handler", "err", err)
 		errutil.ServerErrorResponse(w, r, err)
 		return
 	}
