+++ internal/handler/maincateg/response.go
@@ -1,0 +1,17 @@
+package maincateg
+
+type icon struct {
+	ID  int64  `json:"id"`
+	URL string `json:"url"`
+}
+
+type mainCateg struct {
+	ID   int64  `json:"id"`
+	Name string `json:"name"`
+	Type string `json:"type"`
+	Icon icon   `json:"icon"`
+}
+
+type GetAllMainCategResp struct {
+	Categories []mainCateg `json:"categories"`
+}
